+++ WeChip.Repository/Interface/IStatusRepository.cs
@@ -15,7 +15,7 @@ namespace WeChip.Repository.Interface
         /// </summary>
         /// <param name="statusCode">Código do Status a retornar</param>
         /// <returns></returns>
-        StatusModel Get(byte statusCode);
+        StatusModel Get(short statusCode);
 
     }
 }
